project_2/bytebeaters/requirements.txt
mfrc522==0.0.7
python_bcrypt==0.3.2
rfi==0.1.12
wiringpi==2.60.1
Flask==2.3.2
mysql-connector-python==8.0.33
names==0.3.0
python_bcrypt==0.3.2
cryptography==41.0.0